Reorganize taskworker a bit
