+++ sdks/python/apache_beam/examples/taskworker/local.py
@@ -1,7 +1,23 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 """
-Basic graph to test using TaskWorker.
+Basic pipeline to test using TaskWorkers.
 """
-# pytype: skip-file
 
 from __future__ import absolute_import
 
@@ -31,7 +47,9 @@ class TestFn(beam.DoFn):
 
 
 def run(argv=None, save_main_session=True):
-  """Main entry point; defines and runs the test pipeline."""
+  """
+  Run a pipeline that executes each element using the local task worker.
+  """
   parser = argparse.ArgumentParser()
   known_args, pipeline_args = parser.parse_known_args(argv)
 
@@ -41,7 +59,6 @@ def run(argv=None, save_main_session=True):
   pipeline_options.view_as(SetupOptions).save_main_session = save_main_session
   pipeline_options.view_as(DirectOptions).direct_running_mode = 'multi_processing'
 
-  # The pipeline will be run on exiting the with block.
   with beam.Pipeline(options=pipeline_options) as pipe:
 
     A = (
