+++ sdks/python/build-requirements.txt
@@ -17,4 +17,3 @@
 grpcio-tools==1.30.0
 future==0.18.2
 mypy-protobuf==1.18
-entrypoints==0.3
