+++ sdks/python/setup.py
@@ -141,6 +141,7 @@ REQUIRED_PACKAGES = [
     # server, therefore list of allowed versions is very narrow.
     # See: https://github.com/uqfoundation/dill/issues/341.
     'dill>=0.3.1.1,<0.3.2',
+    'entrypoints>=0.3',
     'fastavro>=0.21.4,<2',
     'funcsigs>=1.0.2,<2; python_version < "3.0"',
     'future>=0.18.2,<1.0.0',
@@ -333,6 +334,9 @@ setuptools.setup(
     entry_points={
         'nose.plugins.0.10': [
             'beam_test_plugin = test_config:BeamTestPlugin',
+        ],
+        'apache_beam_task_workers_plugins': [
+            'k8s_task_worker = apache_beam.task.task_worker.k8s.handler:KubeTaskWorkerHandler',
         ]},
     cmdclass={
         'build_py': generate_protos_first(build_py),
